+++ crates/common/src/util.rs
@@ -12,16 +12,17 @@ use std::{
     sync::{Arc, Mutex},
 };
 
-use crate::error::Error;
-
 #[cfg(feature = "wasm")]
 use wasm_bindgen::prelude::*;
+
+use crate::error::Error;
 
 /// Shared mutable type able to be passed between types
 /// allowing for circular referencing and interior mutability.
 pub type SharedMut<T> = Rc<RefCell<T>>;
 
 /// Shared thread type able to be passed between threads.
+///
 /// Significant performance overhead compared to `SharedMut`.
 pub type SharedThread<T> = Arc<Mutex<T>>;
 
@@ -53,6 +54,7 @@ pub fn write_file(path: &str, data: &[u8], flush: Option<bool>) -> Result<(), Er
 }
 
 /// Replaces the extension in the given path with the provided extension.
+///
 /// This function allows for simple associated file discovery.
 pub fn replace_ext(path: &str, new_extension: &str) -> Option<String> {
     let file_path = Path::new(path);
@@ -147,8 +149,8 @@ pub fn copy_fast(src: &[u8], dst: &mut [u8], count: usize) {
     }
 }
 
-// Interleaves two arrays of bytes into a single array using
-// a pointer-based approach for performance reasons.
+/// Interleaves two arrays of bytes into a single array using
+/// a pointer-based approach for performance reasons.
 pub fn interleave_arrays(a: &[u8], b: &[u8], output: &mut [u8]) {
     assert_eq!(a.len(), b.len());
     assert_eq!(output.len(), a.len() + b.len());
@@ -187,6 +189,10 @@ pub fn flip_vertical(pixels: &[u8], width: usize, height: usize, channels: usize
     flipped
 }
 
+/// Returns the current timestamp in seconds since the UNIX epoch.
+///
+/// This function has different implementations depending on whether
+/// the `wasm` feature is enabled or not.
 #[cfg(not(feature = "wasm"))]
 pub fn timestamp() -> u64 {
     use std::time::{SystemTime, UNIX_EPOCH};
@@ -195,6 +201,12 @@ pub fn timestamp() -> u64 {
     now.duration_since(UNIX_EPOCH).unwrap().as_secs()
 }
 
+/// Returns the current timestamp in seconds since the UNIX epoch.
+///
+/// This function has different implementations depending on whether
+/// the `wasm` feature is enabled or not.
+///
+/// WASM implementation is a static one using `js_sys::Date`.
 #[cfg(feature = "wasm")]
 #[cfg_attr(feature = "wasm", wasm_bindgen)]
 pub fn timestamp() -> u64 {