+++ src/apu.rs
@@ -1521,7 +1521,6 @@ impl Default for Apu {
 #[cfg(test)]
 mod tests {
     use super::{Apu, HighPassFilter};
-
     use crate::state::StateComponent;
 
     #[test]