+++ src/bin/bosc-thumbnail.rs
@@ -5,10 +5,11 @@
 //! name with a `.bmp` extension.
 //!
 //! # Usage
-//! bosc-thumbnail <bosc_file> [thumbnail_file]
+//! bosc-thumbnail <bosc_file> \[thumbnail_file\]
+
+use std::{env::args, error::Error, fs::File, io::Read, path::Path};
 
 use boytacean::state::{SaveStateFormat, StateManager};
-use std::{env::args, error::Error, fs::File, io::Read, path::Path};
 
 fn print_usage() {
     println!("Usage: bosc-thumbnail <bosc_file> [thumbnail_file]");