+++ src/cheats/genie.rs
@@ -1,9 +1,9 @@
-use boytacean_common::error::Error;
 use std::{
     collections::HashMap,
     fmt::{self, Display, Formatter},
 };
 
+use boytacean_common::error::Error;
 #[cfg(feature = "wasm")]
 use wasm_bindgen::prelude::*;